+++ src/public/DeployNovaPackage.ps1
@@ -16,12 +16,17 @@ function Deploy-NovaPackage {
         $uploadOption = New-NovaPackageUploadOption -BoundParameters $PSBoundParameters
         $workflowContext = Get-NovaPackageUploadWorkflowContext -BoundParameters $PSBoundParameters -ProjectInfo $projectInfo -UploadOption $uploadOption
 
+        Write-NovaPackageUploadWorkflowContext -WorkflowContext $workflowContext
+
         $shouldRun = $PSCmdlet.ShouldProcess($workflowContext.Target, $workflowContext.Operation)
 
         if (-not $shouldRun) {
             return @()
         }
 
-        return @(Invoke-NovaPackageUploadWorkflow -WorkflowContext $workflowContext -UploadArtifactList $workflowContext.UploadArtifactList)
+        $result = @(Invoke-NovaPackageUploadWorkflow -WorkflowContext $workflowContext -UploadArtifactList $workflowContext.UploadArtifactList)
+        Write-NovaPackageUploadResultOutput -Result $result
+
+        return $result
     }
 }